// front-end/src/api/models/api/inspirationOption.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Wijckie API
 * Default API for Wijckie
 * OpenAPI spec version: 1.0.0
 */
import type { TypeEnum } from "./typeEnum"

export interface InspirationOption {
    readonly id: number
    readonly module: number
    name: string
    type: TypeEnum
    text?: string
    image?: number
    readonly imageURL: string
}