// front-end/src/api/models/api/typeEnum.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Wijckie API
 * Default API for Wijckie
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `text` - Text
 * `image` - Image
 */
export type TypeEnum = (typeof TypeEnum)[keyof typeof TypeEnum]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TypeEnum = {
    text: "text",
    image: "image",
} as const